fix: correct typo of package name
