+++ src/main/java/nextstep/subway/application/dto/LineRequest.java
@@ -1,9 +1,19 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
-public class SectionRequest {
+public class LineRequest {
+    private String name;
+    private String color;
     private Long upStationId;
     private Long downStationId;
     private int distance;
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
 
     public Long getUpStationId() {
         return upStationId;
