+++ src/main/java/nextstep/subway/application/dto/LineResponse.java
@@ -1,16 +1,21 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
-public class StationResponse {
+public class LineResponse {
     private Long id;
     private String name;
+    private String color;
+    private List<StationResponse> stations;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public LineResponse(Long id, String name, String color, List<StationResponse> stations, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
+        this.color = color;
+        this.stations = stations;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
     }
@@ -23,6 +28,14 @@ public class StationResponse {
         return name;
     }
 
+    public String getColor() {
+        return color;
+    }
+
+    public List<StationResponse> getStations() {
+        return stations;
+    }
+
     public LocalDateTime getCreatedDate() {
         return createdDate;
     }
@@ -31,3 +44,4 @@ public class StationResponse {
         return modifiedDate;
     }
 }
+
