+++ src/main/java/nextstep/subway/application/dto/SectionRequest.java
@@ -1,19 +1,9 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
-public class LineRequest {
-    private String name;
-    private String color;
+public class SectionRequest {
     private Long upStationId;
     private Long downStationId;
     private int distance;
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
 
     public Long getUpStationId() {
         return upStationId;
