+++ src/main/java/nextstep/subway/application/dto/StationResponse.java
@@ -1,21 +1,16 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.application.dto;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
-public class LineResponse {
+public class StationResponse {
     private Long id;
     private String name;
-    private String color;
-    private List<StationResponse> stations;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public LineResponse(Long id, String name, String color, List<StationResponse> stations, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
-        this.color = color;
-        this.stations = stations;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
     }
@@ -28,14 +23,6 @@ public class LineResponse {
         return name;
     }
 
-    public String getColor() {
-        return color;
-    }
-
-    public List<StationResponse> getStations() {
-        return stations;
-    }
-
     public LocalDateTime getCreatedDate() {
         return createdDate;
     }
@@ -44,4 +31,3 @@ public class LineResponse {
         return modifiedDate;
     }
 }
-
